Restructured `sql-update` function into v1 file
